requirements.txt
python-telegram-bot==22.7
pynacl==1.6.2
flask==3.0.0
